DL/Avanzado/Transfer-Style/requirements.txt
Pillow==8.1.2
torch
torchvision
tqdm